+++ src/app/(auth)/signup/_components/signup-form/index.ts
@@ -1,2 +1,3 @@
 export * from './signup-form';
+export * from './signup-form.schema';
 export * from './signup-form.types';
